backend/requirements.txt
annotated-doc
annotated-types
anyio
certifi
charset-normalizer
click
colorama
distro
fastapi
filelock
fsspec
h11
httpcore
httptools
httpx
huggingface-hub
idna
Jinja2
jiter
joblib
lxml
MarkupSafe
mpmath
numpy
openai
packaging
pillow
pydantic
pydantic_core
pypdf
python-docx
python-dotenv
python-multipart
python-pptx
PyYAML
regex
requests
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
sniffio
starlette
sympy
threadpoolctl
tokenizers
torch
tqdm
transformers
typing-inspection
typing_extensions
urllib3
uvicorn
watchfiles
websockets
xlsxwriter
